Function comp_sci::remove_duplicates_by_sorting

pub fn remove_duplicates_by_sorting<'a, A: PartialEq + Ord>(data: &'a mut Vec<A>)

Removes duplicate entries from Vec with a complexity of O(n log n + n) I believe (TODO).

This technique sorts the vector before removing the duplicates and thus is not stable.